+++ src/Files.App/ViewModels/UserControls/AddressToolbarViewModel.cs
@@ -3,63 +3,29 @@
 
 using CommunityToolkit.WinUI.UI;
 using Files.Shared.Helpers;
-using Microsoft.UI.Dispatching;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Controls.Primitives;
 using Microsoft.UI.Xaml.Input;
 using System.IO;
 using System.Windows.Input;
-using Windows.ApplicationModel.DataTransfer;
-using Windows.UI.Text;
 
 namespace Files.App.ViewModels.UserControls
 {
-	public class ToolbarViewModel : ObservableObject, IAddressToolbar, IDisposable
+	public class AddressToolbarViewModel : ObservableObject, IAddressToolbar, IDisposable
 	{
 		private IUserSettingsService UserSettingsService { get; } = Ioc.Default.GetRequiredService<IUserSettingsService>();
-
 		private readonly IDialogService _dialogService = Ioc.Default.GetRequiredService<IDialogService>();
-
-		private readonly MainPageViewModel mainPageViewModel = Ioc.Default.GetRequiredService<MainPageViewModel>();
-
 		private readonly DrivesViewModel drivesViewModel = Ioc.Default.GetRequiredService<DrivesViewModel>();
-
 		public IUpdateService UpdateService { get; } = Ioc.Default.GetService<IUpdateService>()!;
-
 		public ICommandManager Commands { get; } = Ioc.Default.GetRequiredService<ICommandManager>();
 
-		public delegate void ToolbarPathItemInvokedEventHandler(object sender, PathNavigationEventArgs e);
-
-		public delegate void ToolbarFlyoutOpenedEventHandler(object sender, ToolbarFlyoutOpenedEventArgs e);
-
-		public delegate void ToolbarPathItemLoadedEventHandler(object sender, ToolbarPathItemLoadedEventArgs e);
-
 		public delegate void AddressBarTextEnteredEventHandler(object sender, AddressBarTextEnteredEventArgs e);
-
-		public delegate void PathBoxItemDroppedEventHandler(object sender, PathBoxItemDroppedEventArgs e);
-
-		public event ToolbarPathItemInvokedEventHandler? ToolbarPathItemInvoked;
-
-		public event ToolbarFlyoutOpenedEventHandler? ToolbarFlyoutOpened;
-
-		public event ToolbarPathItemLoadedEventHandler? ToolbarPathItemLoaded;
-
-		public event IAddressToolbar.ItemDraggedOverPathItemEventHandler? ItemDraggedOverPathItem;
-
 		public event EventHandler? EditModeEnabled;
-
 		public event IAddressToolbar.ToolbarQuerySubmittedEventHandler? PathBoxQuerySubmitted;
-
 		public event AddressBarTextEnteredEventHandler? AddressBarTextEntered;
-
-		public event PathBoxItemDroppedEventHandler? PathBoxItemDropped;
-
 		public event EventHandler? RefreshRequested;
-
 		public event EventHandler? RefreshWidgetsRequested;
-
-		public ObservableCollection<PathBoxItem> PathComponents { get; } = new();
 
 		private bool _isCommandPaletteOpen;
 		public bool IsCommandPaletteOpen
@@ -199,15 +165,10 @@ namespace Files.App.ViewModels.UserControls
 			}
 		}
 
-		private PointerRoutedEventArgs? pointerRoutedEventArgs;
-
-		public ToolbarViewModel()
+		public AddressToolbarViewModel()
 		{
 			RefreshClickCommand = new RelayCommand<RoutedEventArgs>(e => RefreshRequested?.Invoke(this, EventArgs.Empty));
 			ViewReleaseNotesAsyncCommand = new AsyncRelayCommand(ViewReleaseNotesAsync);
-
-			dispatcherQueue = DispatcherQueue.GetForCurrentThread();
-			dragOverTimer = dispatcherQueue.CreateTimer();
 
 			SearchBox.Escaped += SearchRegion_Escaped;
 			UserSettingsService.OnSettingChangedEvent += UserSettingsService_OnSettingChangedEvent;
@@ -265,9 +226,6 @@ namespace Files.App.ViewModels.UserControls
 			}
 		}
 
-		private DispatcherQueue dispatcherQueue;
-		private DispatcherQueueTimer dragOverTimer;
-
 		private ISearchBox searchBox = new SearchBoxViewModel();
 		public ISearchBox SearchBox
 		{
@@ -277,134 +235,6 @@ namespace Files.App.ViewModels.UserControls
 
 		public SearchBoxViewModel SearchBoxViewModel
 			=> (SearchBoxViewModel)SearchBox;
-
-		public bool IsSingleItemOverride { get; set; } = false;
-
-		private string? dragOverPath = null;
-
-		public void PathBoxItem_DragLeave(object sender, DragEventArgs e)
-		{
-			if (((StackPanel)sender).DataContext is not PathBoxItem pathBoxItem ||
-				pathBoxItem.Path == "Home")
-			{
-				return;
-			}
-
-			// Reset dragged over pathbox item
-			if (pathBoxItem.Path == dragOverPath)
-				dragOverPath = null;
-		}
-
-		private bool lockFlag = false;
-
-		public async Task PathBoxItem_Drop(object sender, DragEventArgs e)
-		{
-			if (lockFlag)
-				return;
-
-			lockFlag = true;
-
-			// Reset dragged over pathbox item
-			dragOverPath = null;
-
-			if (((StackPanel)sender).DataContext is not PathBoxItem pathBoxItem ||
-				pathBoxItem.Path == "Home")
-			{
-				return;
-			}
-
-			var deferral = e.GetDeferral();
-
-			var signal = new AsyncManualResetEvent();
-
-			PathBoxItemDropped?.Invoke(this, new PathBoxItemDroppedEventArgs()
-			{
-				AcceptedOperation = e.AcceptedOperation,
-				Package = e.DataView,
-				Path = pathBoxItem.Path,
-				SignalEvent = signal
-			});
-
-			await signal.WaitAsync();
-
-			deferral.Complete();
-			await Task.Yield();
-
-			lockFlag = false;
-		}
-
-		public async Task PathBoxItem_DragOver(object sender, DragEventArgs e)
-		{
-			if (IsSingleItemOverride ||
-				((StackPanel)sender).DataContext is not PathBoxItem pathBoxItem ||
-				pathBoxItem.Path == "Home")
-			{
-				return;
-			}
-
-			if (dragOverPath != pathBoxItem.Path)
-			{
-				dragOverPath = pathBoxItem.Path;
-				dragOverTimer.Stop();
-
-				if (dragOverPath != (this as IAddressToolbar).PathComponents.LastOrDefault()?.Path)
-				{
-					dragOverTimer.Debounce(() =>
-					{
-						if (dragOverPath is not null)
-						{
-							dragOverTimer.Stop();
-							ItemDraggedOverPathItem?.Invoke(this, new PathNavigationEventArgs()
-							{
-								ItemPath = dragOverPath
-							});
-							dragOverPath = null;
-						}
-					},
-					TimeSpan.FromMilliseconds(1000), false);
-				}
-			}
-
-			// In search page
-			if (!FilesystemHelpers.HasDraggedStorageItems(e.DataView) || string.IsNullOrEmpty(pathBoxItem.Path))
-			{
-				e.AcceptedOperation = DataPackageOperation.None;
-
-				return;
-			}
-
-			e.Handled = true;
-			var deferral = e.GetDeferral();
-
-			var storageItems = await FilesystemHelpers.GetDraggedStorageItems(e.DataView);
-
-			if (!storageItems.Any(storageItem =>
-					!string.IsNullOrEmpty(storageItem?.Path) &&
-					storageItem.Path.Replace(pathBoxItem.Path, string.Empty, StringComparison.Ordinal)
-						.Trim(Path.DirectorySeparatorChar)
-						.Contains(Path.DirectorySeparatorChar)))
-			{
-				e.AcceptedOperation = DataPackageOperation.None;
-			}
-
-			// Copy be default when dragging from zip
-			else if (storageItems.Any(x =>
-					x.Item is ZipStorageFile ||
-					x.Item is ZipStorageFolder) ||
-					ZipStorageFolder.IsZipPath(pathBoxItem.Path))
-			{
-				e.DragUIOverride.Caption = string.Format("CopyToFolderCaptionText".GetLocalizedResource(), pathBoxItem.Title);
-				e.AcceptedOperation = DataPackageOperation.Copy;
-			}
-			else
-			{
-				e.DragUIOverride.IsCaptionVisible = true;
-				e.DragUIOverride.Caption = string.Format("MoveToFolderCaptionText".GetLocalizedResource(), pathBoxItem.Title);
-				e.AcceptedOperation = DataPackageOperation.Move;
-			}
-
-			deferral.Complete();
-		}
 
 		public bool IsEditModeEnabled
 		{
@@ -454,24 +284,6 @@ namespace Files.App.ViewModels.UserControls
 		public ICommand RefreshClickCommand { get; }
 		public ICommand ViewReleaseNotesAsyncCommand { get; }
 
-		public void PathItemSeparator_DataContextChanged(FrameworkElement sender, DataContextChangedEventArgs args)
-		{
-			var pathSeparatorIcon = sender as FontIcon;
-			if (pathSeparatorIcon is null || pathSeparatorIcon.DataContext is null)
-				return;
-
-			ToolbarPathItemLoaded?.Invoke(pathSeparatorIcon, new ToolbarPathItemLoadedEventArgs()
-			{
-				Item = (PathBoxItem)pathSeparatorIcon.DataContext,
-				OpenedFlyout = (MenuFlyout)pathSeparatorIcon.ContextFlyout
-			});
-		}
-
-		public void PathboxItemFlyout_Opened(object sender, object e)
-		{
-			ToolbarFlyoutOpened?.Invoke(this, new ToolbarFlyoutOpenedEventArgs() { OpenedFlyout = (MenuFlyout)sender });
-		}
-
 		public void VisiblePath_TextChanged(AutoSuggestBox sender, AutoSuggestBoxTextChangedEventArgs args)
 		{
 			if (args.Reason == AutoSuggestionBoxTextChangeReason.UserInput)
@@ -483,39 +295,6 @@ namespace Files.App.ViewModels.UserControls
 			PathBoxQuerySubmitted?.Invoke(this, new ToolbarQuerySubmittedEventArgs() { QueryText = args.QueryText });
 
 			(this as IAddressToolbar).IsEditModeEnabled = false;
-		}
-
-		public void PathBoxItem_PointerPressed(object sender, PointerRoutedEventArgs e)
-		{
-			if (e.Pointer.PointerDeviceType != Microsoft.UI.Input.PointerDeviceType.Mouse)
-				return;
-
-			var ptrPt = e.GetCurrentPoint(AddressToolbar);
-			pointerRoutedEventArgs = ptrPt.Properties.IsMiddleButtonPressed ? e : null;
-		}
-
-		public async Task PathBoxItem_Tapped(object sender, TappedRoutedEventArgs e)
-		{
-			var itemTappedPath = ((sender as TextBlock)?.DataContext as PathBoxItem)?.Path;
-			if (itemTappedPath is null)
-				return;
-
-			if (pointerRoutedEventArgs is not null)
-			{
-				await MainWindow.Instance.DispatcherQueue.EnqueueOrInvokeAsync(async () =>
-				{
-					await mainPageViewModel.AddNewTabByPathAsync(typeof(PaneHolderPage), itemTappedPath);
-				}, DispatcherQueuePriority.Low);
-				e.Handled = true;
-				pointerRoutedEventArgs = null;
-
-				return;
-			}
-
-			ToolbarPathItemInvoked?.Invoke(this, new PathNavigationEventArgs()
-			{
-				ItemPath = itemTappedPath
-			});
 		}
 
 		public void OpenCommandPalette()
@@ -604,67 +383,6 @@ namespace Files.App.ViewModels.UserControls
 		public ICommand? Share { get; set; }
 
 		public ICommand? UpdateCommand { get; set; }
-
-		public async Task SetPathBoxDropDownFlyoutAsync(MenuFlyout flyout, PathBoxItem pathItem, IShellPage shellPage)
-		{
-			var nextPathItemTitle = PathComponents[PathComponents.IndexOf(pathItem) + 1].Title;
-			IList<StorageFolderWithPath>? childFolders = null;
-
-			StorageFolderWithPath folder = await shellPage.FilesystemViewModel.GetFolderWithPathFromPathAsync(pathItem.Path);
-			if (folder is not null)
-				childFolders = (await FilesystemTasks.Wrap(() => folder.GetFoldersWithPathAsync(string.Empty))).Result;
-
-			flyout.Items?.Clear();
-
-			if (childFolders is null || childFolders.Count == 0)
-			{
-				var flyoutItem = new MenuFlyoutItem
-				{
-					Icon = new FontIcon { Glyph = "\uE7BA" },
-					Text = "SubDirectoryAccessDenied".GetLocalizedResource(),
-					//Foreground = (SolidColorBrush)Application.Current.Resources["SystemControlErrorTextForegroundBrush"],
-					FontSize = 12
-				};
-
-				flyout.Items?.Add(flyoutItem);
-
-				return;
-			}
-
-			var boldFontWeight = new FontWeight { Weight = 800 };
-			var normalFontWeight = new FontWeight { Weight = 400 };
-
-			var workingPath =
-				PathComponents[PathComponents.Count - 1].Path?.TrimEnd(Path.DirectorySeparatorChar);
-
-			foreach (var childFolder in childFolders)
-			{
-				var isPathItemFocused = childFolder.Item.Name == nextPathItemTitle;
-
-				var flyoutItem = new MenuFlyoutItem
-				{
-					Icon = new FontIcon
-					{
-						Glyph = "\uED25",
-						FontWeight = isPathItemFocused ? boldFontWeight : normalFontWeight
-					},
-					Text = childFolder.Item.Name,
-					FontSize = 12,
-					FontWeight = isPathItemFocused ? boldFontWeight : normalFontWeight
-				};
-
-				if (workingPath != childFolder.Path)
-				{
-					flyoutItem.Click += (sender, args) =>
-					{
-						// Navigate to the directory
-						shellPage.NavigateToPath(childFolder.Path);
-					};
-				}
-
-				flyout.Items?.Add(flyoutItem);
-			}
-		}
 
 		public async Task CheckPathInputAsync(string currentInput, string currentSelectedPath, IShellPage shellPage)
 		{
