+++ src/Files.App/ViewModels/UserControls/IAddressToolbar.cs
@@ -1,11 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.Views;
-using Microsoft.UI.Xaml.Controls;
-using Windows.ApplicationModel.DataTransfer;
-
-namespace Files.App.UserControls
+namespace Files.App.ViewModels.UserControls
 {
 	public interface IAddressToolbar
 	{
@@ -13,9 +9,6 @@ namespace Files.App.UserControls
 
 		public bool IsEditModeEnabled { get; set; }
 
-		/// <summary>
-		/// Boolean to determine if the command palette is open
-		/// </summary>
 		public bool IsCommandPaletteOpen { get; set; }
 
 		public bool CanRefresh { get; set; }
@@ -28,21 +21,13 @@ namespace Files.App.UserControls
 
 		public bool CanGoForward { get; set; }
 
-		public bool IsSingleItemOverride { get; set; }
-
 		public string PathControlDisplayText { get; set; }
 
-		public ObservableCollection<PathBoxItem> PathComponents { get; }
-
 		public delegate void ToolbarQuerySubmittedEventHandler(object sender, ToolbarQuerySubmittedEventArgs e);
-
-		public delegate void ItemDraggedOverPathItemEventHandler(object sender, PathNavigationEventArgs e);
 
 		public event ToolbarQuerySubmittedEventHandler PathBoxQuerySubmitted;
 
 		public event EventHandler EditModeEnabled;
-
-		public event ItemDraggedOverPathItemEventHandler ItemDraggedOverPathItem;
 
 		public event EventHandler RefreshRequested;
