+++ src/Files.App/Views/Shells/BaseShellPage.cs
@@ -49,7 +49,7 @@ namespace Files.App.Views.Shells
 
 		protected readonly ICommandManager commands = Ioc.Default.GetRequiredService<ICommandManager>();
 
-		public ToolbarViewModel ToolbarViewModel { get; } = new ToolbarViewModel();
+		public AddressToolbarViewModel ToolbarViewModel { get; } = new AddressToolbarViewModel();
 
 		public IBaseLayout SlimContentPage => ContentPage;
